refactor: extract sub header
